requirements.txt
charset-normalizer==2.0.6
Flask==2.0.1
Jinja2==3.0.1
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.0.3
MarkupSafe==2.0.1
PyMySQL==1.0.2
Werkzeug==2.0.1
PyGithub==1.55
python-decouple==3.5
cryptography==36.0.1
